requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.30
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
narwhals==1.34.1
numpy==2.2.4
openai==1.73.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.44.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.19.0
zstandard==0.23.0
